refactor obstacle nav, use lat,lng order
