+++ web/frontend/src/components/remote-control-cards.svelte
@@ -167,7 +167,7 @@ const getStatus = (statusMap: Record<string, unknown>, resource: commonApi.Resou
 
     <!-- ******* NAVIGATION *******  -->
     {#each filterSubtype($services, 'navigation') as { name } (name)}
-      <Navigation {name} write={false} />
+      <Navigation {name} write={true} />
     {/each}
 
     <!-- ******* SENSOR *******  -->
